pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "file_chopper"
version = "1.0.0"
description = "Split large files into smaller pieces and rejoin them."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
keywords = ["file", "split", "chunks", "join", "large files"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "tqdm>=4.64",
    "pypdf>=4.0",
    "python-docx>=1.0",
    "python-pptx>=0.6",
    "openpyxl>=3.1",
    "odfpy>=1.4",
    "striprtf>=0.0.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[project.scripts]
file_chopper = "file_chopper.main:cli"

[project.urls]
Homepage = "https://github.com/Buttje/file_chopper"
"Bug Tracker" = "https://github.com/Buttje/file_chopper/issues"

[tool.setuptools.packages.find]
where = ["src"]

# ---------------------------------------------------------------------------
# Ruff (linter + formatter)
# ---------------------------------------------------------------------------

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # allow assert statements in tests

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
